refactor: lift tree row into ui::tree::item — Kyde-agnostic component
tree_row (an impl Kyde method) becomes a free generic-over-view component in
src/ui/tree.rs: knows nothing about Kyde, takes row data + 3 handlers +
suppress_hover. The 4 callers (browse/commit/history/rollback) pass
self.dragging(Divider::Tree) for hover suppression. Zero behaviour change.
